Spring data capture!

+++ joytest/joytest.c
@@ -83,7 +83,7 @@ void VendorRequests(void) {
             BD[EP0IN].status = 0xC8;
             break;
         case JOY_READ_D:
-            result32 = (WORD32)((float)joy.cmd/65535.0);
+            result32 = (WORD32)(float)((float)joy.cmd/65535.0);
             BD[EP0IN].address[0] = result32.b[0];
             BD[EP0IN].address[1] = result32.b[1];
             BD[EP0IN].address[2] = result32.b[2];
